Glob finished

+++ src/GameObject.java
@@ -1,10 +1,10 @@
 public class GameObject{
-	int x;
-	int y;
+	double x;
+	double y;
 	int width;
 	int height;
 	
-	public GameObject(int x, int y, int width, int height) {
+	public GameObject(double x, double y, int width, int height) {
 		this.x = x;
 		this.y = y;
 		this.width = width;
